+++ server/controllers/PurchaseItemsController.ts
@@ -4,9 +4,14 @@ import IPurchaseItemsService from "../services/interfaces/PurchaseItemsService";
 import PurchaseItemsService from "../services/PurchaseItemsService";
 import InsertPurchaseItemResponse from "../models/responses/InsertPurchaseItemsResponse";
 import GetPurchaseItemsResponse from "../models/responses/GetPurchaseItemsResponse";
+import MapHelper from "../common/MapHelper";
+import AirtableService from "../services/AirtableService";
 
 const purhaseItemsController = express.Router();
-const purchaseItemsService: IPurchaseItemsService = new PurchaseItemsService();
+const purchaseItemsService: IPurchaseItemsService = new PurchaseItemsService(
+    new AirtableService(),
+    new MapHelper()
+);
 
 purhaseItemsController.route("/getPurchaseItems").post(
     async (req, res): Promise<void> => {
