+++ server/services/PurchaseItemsService.ts
@@ -1,7 +1,6 @@
 import IPurchaseItemsService from "./interfaces/PurchaseItemsService";
-import { AirtableInsertRequest } from "./interfaces/AirtableService";
+import IMapHelper from "../common/interfaces/MapHelper";
 import PurchaseItem from "../models/PurchanseItem";
-import AirtableService from "./AirtableService";
 import IAirtableService from "./interfaces/AirtableService";
 import * as Moment from "moment";
 import InsertPurchaseItemsRequest from "../models/requests/InsertPurchaseItemsRequest";
@@ -9,9 +8,14 @@ import GetPurchaseItemsRequest from "../models/requests/GetPurchaseItemsRequest"
 
 class PurchaseItemsService implements IPurchaseItemsService {
     private airtableService: IAirtableService;
+    private mapHelper: IMapHelper;
 
-    public constructor() {
-        this.airtableService = new AirtableService();
+    public constructor(
+        airtableService: IAirtableService,
+        mapHelper: IMapHelper
+    ) {
+        this.airtableService = airtableService;
+        this.mapHelper = mapHelper;
         Moment.locale("th");
     }
 
@@ -47,7 +51,9 @@ class PurchaseItemsService implements IPurchaseItemsService {
                                 if (page === 0) {
                                     toReturn = records.map(
                                         (record: any): PurchaseItem =>
-                                            this.mapPurchaseItem(record)
+                                            this.mapHelper.mapAirtableRecToPurchaseItem(
+                                                record
+                                            )
                                     );
                                     resolve(toReturn);
                                 }
@@ -59,7 +65,9 @@ class PurchaseItemsService implements IPurchaseItemsService {
 
                         toReturn = records.map(
                             (record: any): PurchaseItem =>
-                                this.mapPurchaseItem(record)
+                                this.mapHelper.mapAirtableRecToPurchaseItem(
+                                    record
+                                )
                         );
                         resolve(toReturn);
                     }
@@ -75,7 +83,7 @@ class PurchaseItemsService implements IPurchaseItemsService {
         request: InsertPurchaseItemsRequest
     ): Promise<void> {
         try {
-            const inserReq = this.mapRequestInsertPurchaseItem(
+            const inserReq = this.mapHelper.mapRequestToPurchaseItem(
                 request.purchaseItems
             );
 
@@ -83,58 +91,6 @@ class PurchaseItemsService implements IPurchaseItemsService {
         } catch (err) {
             console.log(err);
 
-            throw err;
-        }
-    }
-
-    private mapPurchaseItem(record: any): PurchaseItem {
-        try {
-            const toReturn = {
-                purchaseId: record.fields["Purchase ID"],
-                paymentDueDate: Moment(
-                    record.fields["Payment Due Date"],
-                    "D MMMM YYYY"
-                ).toDate(),
-                account: record.fields["Account"],
-                thbInvoiceAmount: record.fields["THB Invoice Amount"],
-                category: record.fields["Category"],
-                enteredBy: record.fields["Entered By"],
-                requestBy: record.fields["Request By"],
-                createdTime: Moment(
-                    record.fields["Created Time"],
-                    "D MMMM YYYY h:mma"
-                ).toDate()
-            };
-
-            return toReturn as PurchaseItem;
-        } catch (err) {
-            throw err;
-        }
-    }
-
-    private mapRequestInsertPurchaseItem(
-        request: PurchaseItem[]
-    ): AirtableInsertRequest<PurchaseItem>[] {
-        try {
-            return request.map(x => {
-                const toReturn = {};
-                toReturn["Purchase ID"] = x.purchaseId;
-                toReturn["Payment Due Date"] = x.paymentDueDate;
-                toReturn["Account"] = x.account;
-                toReturn["THB Invoice Amount"] = x.thbInvoiceAmount;
-                toReturn["Category"] = x.category;
-                toReturn["Entered By"] = x.enteredBy;
-                toReturn["Request By"] = x.requestBy;
-                toReturn["Created Time"] = Moment(
-                    Moment.now(),
-                    "D MMMM YYYY h:mma"
-                ).toString();
-
-                return {
-                    fields: toReturn
-                };
-            }) as AirtableInsertRequest<PurchaseItem>[];
-        } catch (err) {
             throw err;
         }
     }
